Experiment
Currently version of acpcdetect command will apply ACPC alignment to the voxel and leave sform to be identical (with translation to be almost the middle of the voxel space). I'd like to try leaving the voxel to be untouched but let it update the sform instead - so that we can preserve information needed to go from voxel to the original scanner space. This change *shouldn't* affect existing app (!)

Also fixed the "crop_reorient" key.